fix: move regular drop banner to home
The vote banner was mounted on RewardsScreen, but the intended placement
"directly under Hi {name}" lives on HomeScreen (GreetingHeader). Move the
banner + shared RegularDropSheet mount to HomeScreen and restore
RewardsScreen to its pre-banner state. Same components, same user-scoped
query cache — no new logic, no extra fetch, no backend change.
